requirements.txt
# All required Python packages will be listed here.

python-telegram-bot
fastapi
pymongo
apscheduler
python-dotenv
google-generativeai
python-telegram-bot[job-queue]